[PDF Ink Signatures] Use uint32_t in PdfCaretClient::GetCharCount()
In https://crrev.com/c/6720352, PdfCaret will use PageCharacterIndex
structs, which uses unsigned ints. To avoid having to statically cast
the return values in PdfCaretClient::GetCharCount() to compare ints, use
unsigned ints in GetCharCount(), too.

Also make the page index an unsigned int, since negative page indices
are invalid and GetCharCount() crashes anyway for invalid page indices.

Bug: 427133561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6735098
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487231}

+++ pdf/pdfium/pdfium_engine.cc
@@ -943,9 +943,9 @@ void PDFiumEngine::OnDocumentCanceled() {
     OnDocumentComplete();
 }
 
-int PDFiumEngine::GetCharCount(int page_index) const {
+uint32_t PDFiumEngine::GetCharCount(uint32_t page_index) const {
   CHECK(PageIndexInBounds(page_index));
-  return pages_[page_index]->GetCharCount();
+  return base::checked_cast<uint32_t>(pages_[page_index]->GetCharCount());
 }
 
 std::vector<gfx::Rect> PDFiumEngine::GetScreenRectsForChar(
